requirements.txt
Django==1.6.6
Pillow==2.5.3
PyYAML==3.11
argparse==1.2.1
boto==2.32.1
dj-database-url==0.3.0
dj-static==0.0.6
django-athumb==2.0
django-bower==5.0.1
django-localflavor==1.0
django-sampledatahelper==0.2.1
django-toolbelt==0.0.1
django-widget-tweaks==1.3
gunicorn==19.1.1
psycopg2==2.5.3
pystache==0.5.4
requests==2.3.0
six==1.7.3
static==1.0.2
static3==0.5.1
stripe==1.19.0
wsgiref==0.1.2
